#155 Reverting it_poster_test and test_poster modules
